+++ mobitopp/src/main/java/edu/kit/ifv/mobitopp/simulation/activityschedule/ActivityIfc.java
@@ -16,7 +16,7 @@ public interface ActivityIfc
   boolean isRunning();
   void setRunning(boolean running_);
 
-  int getActivityNrOfWeek();
+  byte getActivityNrOfWeek();
  
 	ActivityType activityType();
